refactor: overhaul backend architecture, logging, and tests
- Refactor be/main.py to focus on bootstrapping
- Move game logic to be/routers/games.py
- Move rate limits to be/dependencies.py
- Remove psycopg2-binary in favor of asyncpg
- Disable verbose SQLAlchemy logging
- Suppress uvicorn access logs
- Fix test suite imports and mocking strategies
- Add be/pyproject.toml for pytest configuration
