+++ be/requirements.txt
@@ -29,4 +29,4 @@ starlette==0.52.1
 typing-inspection==0.4.2
 typing_extensions==4.15.0
 uvicorn==0.40.0
-psycopg2-binary==2.9.9
+
