+++ be/schemas.py
@@ -1,4 +1,4 @@
-from pydantic import BaseModel, field_validator
+from pydantic import BaseModel, field_validator, ConfigDict
 from typing import List, Optional, Union
 import string
 
@@ -64,5 +64,4 @@ class GameState(BaseModel):
     draw_offer: Optional[str] = None
     draw_timer: Optional[int] = 0
 
-    class Config:
-        from_attributes = True
+    model_config = ConfigDict(from_attributes=True)
